Completed Branch take home project
Completed todo list project which supports

- Creating a new list
- Deleting a list
- Searching for a list based on name or description
- Creating a todo list item
- Deleting a todo list item
- Marking a todo list item as completed
- Searching for a specific todo item within a list

Written using SwiftUI and support iPhone and iPad interfaces. This
project uses the latest iOS15 SDK features so the minimum deployment
target is iOS 15.Uses MVVM architecture and Core Data for data persistance.

If I were to continue with this project, I would like to allow the user to
select a custom image for their list along with a custom accent color
for each list similar to the reminders app. I would also like to allow
the user to see their completed todo items for each list.